Release/v1.1 (#14)
* Change to next version.  Add develop branch to github actions (#8)

* Add playwright 1.19.0 (#9)

* Added ability to define custom field locators (#12)

* Add ability to call a method automatically after creation of page object (#13)

* Prepare v1.1 release

* Update readme with new afterCreate functionality
