+++ .github/workflows/maven.yml
@@ -2,9 +2,9 @@ name: Stagehand
 
 on:
   push:
-    branches: [ main ]
+    branches: [ main, develop ]
   pull_request:
-    branches: [ main ]
+    branches: [ main, develop ]
 
 jobs:
   build:
@@ -12,7 +12,7 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         java: [ '8', '11', '17' ]
-        playwright: ['1.18.0']
+        playwright: ['1.18.0', '1.19.0', '1.20.1']
     runs-on: ${{ matrix.os }}
     name: ${{matrix.os}} - Java ${{ matrix.java }} - Playwright ${{matrix.playwright}}
     steps:
@@ -22,4 +22,4 @@ jobs:
         with:
           distribution: 'temurin'
           java-version: ${{ matrix.java }}
-      - run: mvn test "-Dplaywright.version=${{matrix.playwright}}" --file pom.xml
+      - run: mvn clean test "-Dplaywright.version=${{matrix.playwright}}" --file pom.xml
