+++ src/main/java/io/github/uchagani/stagehand/AfterCreate.java
@@ -1,0 +1,5 @@
+package io.github.uchagani.stagehand;
+
+public interface AfterCreate {
+    void afterCreate();
+}
